refactor: address code review — u32 indices, AssetStore, Scale, test wiring
- Separate query_generation from current_frame to fix cache invalidation
- Switch from u16 to u32 indices (supports meshes >65K vertices)
- Wire tests.zig integration tests into zig build test (46 tests total)
- Extract AssetStore struct from Engine for mesh/material/texture registries
- Add Scale component with auto Lua bindings, used in renderer model matrix
- Add Mat4.scale() to math3d.zig
- Give MeshHandle, MaterialHandle, LookAt proper Lua bindings (remove ~40 LOC)
- Expose metallic, roughness, emissive in luaCreateMaterial
- Pass dt to renderer.renderSystem for future GPU-side animation
- Fix cube and sphere tangent generation for correct normal mapping
- Eliminate double-build of pbr_test in build.zig default run step
- Add TODO for hardcoded homebrew paths

+++ src/gltf.zig
@@ -182,12 +182,12 @@ fn loadPrimitive(engine: *Engine, allocator: std.mem.Allocator, prim: *const cgl
     }
 
     // Read indices
-    var indices: ?[]u16 = null;
+    var indices: ?[]u32 = null;
     defer if (indices) |idx| allocator.free(idx);
 
     if (prim.indices) |idx_acc_ptr| {
         const idx_count = idx_acc_ptr.*.count;
-        var idx = try allocator.alloc(u16, idx_count);
+        var idx = try allocator.alloc(u32, idx_count);
         for (0..idx_count) |ii| {
             idx[ii] = @intCast(cgltf.cgltf_accessor_read_index(idx_acc_ptr, ii));
         }
